Grant leader lease to ranges; implement HasLeaderLease.
Added new Lease type to proto/data.proto. Leases contain an expiration
as a wall time in unix nanos as well as the node and store IDs of the
replica being granted the lease. Ranges keep track of the lease and can
now answer the question HasLeaderLease() for a range. Further, the last
set lease is used to construct proper NotLeaderError error returns.

Timestamp cache and command cache are cleared on any gap in the lease
holder's tenure.

This CL is still a WIP; needs many fixes to existing unittests.