auto-election for single-node groups
whenever MultiRaft creates a new group, and that group has
only the local node as its member, automatically call elections.

added a helper to the to wait for an election event.

adjusted all tests; had to make minor adjustments in some cases,
and one test is disabled since it relied on the absence of HardState
updates. The minor adjustments are usually waiting out the leader
election event, as otherwise tests could wind up flaky, as in #702,
or due to outstanding work on the mechanism of re-proposing Raft
commands (currently, Raft does that, which means commands may by-
pass the response cache and cause test failures).

the KV test now runs in <5sec on my machine; that's a 100% speedup.

upped the tick interval during tests to 100ms, so in reality we do
not seem to rely on ticks anywhere at the moment.
Of course, that also suggests that we'll need to come up with quite
a bunch of new tests to thoroughly test leader leases, which are due
to land soon.